+++ ProjectSalesCore/ProjectSalesCore.DataBase/Models/Bank.cs
@@ -19,7 +19,7 @@ namespace CSales.Database.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("IDBANK")]
-        public long IdBank { get; set; }
+        public int IdBank { get; set; }
 
         [Column("BANKNAME")]
         public string BankName { get; set; }
