+++ ProjectSalesCore/ProjectSalesCore.DataBase/Models/CurrentAccountDocumentType.cs
@@ -15,7 +15,7 @@
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("ID")]
-        public long IdCurrentAccountDocumentType { get; set; }
+        public int IdCurrentAccountDocumentType { get; set; }
 
         [Column("TYPENAME")]
         public string TypeName { get; set; }
