+++ ProjectSalesCore/ProjectSalesCore.DataBase/Models/DebitNote.cs
@@ -19,31 +19,31 @@ namespace CSales.Database.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("ID")]
-        public long IdDebitNote { get; set; }
+        public int IdDebitNote { get; set; }
 
         [ForeignKey(nameof(Client))]
         [Column("IDCLIENT")]
-        public long IdClient { get; set; }
+        public int IdClient { get; set; }
 
         [Column("CLIENT")]
         public virtual Client Client { get; set; }
 
         [ForeignKey(nameof(Employee))]
         [Column("IDEMPLOYEE")]
-        public long IdEmployee { get; set; }
+        public int IdEmployee { get; set; }
 
         [Column("EMPLOYEE")]
         public virtual Employee Employee { get; set; }
 
         [ForeignKey(nameof(TOSDoc))]
         [Column("IDDOCFORSALE")]
-        public long IdTOSDoc { get; set; }
+        public int IdTOSDoc { get; set; }
 
         [Column("IDTDFORSALE")]
         public virtual TOSDoc TOSDoc { get; set; }
 
         [Column("IDDOCUMENT")]
-        public long IdDocument { get; set; }
+        public int IdDocument { get; set; }
 
         [Column("TOTAL")]
         public decimal Total { get; set; }
