+++ ProjectSalesCore/ProjectSalesCore.DataBase/Models/EmailProvider.cs
@@ -19,14 +19,14 @@ namespace ProjectSalesCore.DataBase.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("IDEMAIL")]
-        public long IdEmail { get; set; }
+        public int IdEmail { get; set; }
 
         [Column("EMAILL")]
         public string Emaill { get; set; }
 
         [ForeignKey(nameof(Provider))]
         [Column("IDPRV")]
-        public long IdPRV { get; set; }
+        public int IdPRV { get; set; }
 
         [Column("PRV")]
         public virtual Provider Provider { get; set; }
