+++ ProjectSalesCore/ProjectSalesCore.DataBase/Models/TelephoneEmployee.cs
@@ -14,7 +14,7 @@ namespace ProjectSalesCore.DataBase.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("ID")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Column("NUMBER")]
         public string Number { get; set; }
@@ -24,7 +24,7 @@ namespace ProjectSalesCore.DataBase.Models
 
         [ForeignKey(nameof(Employee))]
         [Column("IDEMP")]
-        public long IdEMP { get; set; }
+        public int IdEMP { get; set; }
 
         [Column("EMP")]
         public virtual Employee Employee { get; set; }
